Re-throttle window frame callbacks
This was unthrottled because it would mess with snapshots. Evidently the problem that caused me to implement the hack is now fixed (or seems to be, at least), so we can just slide the stuff back where it was. Also hey, turns out if you don't throttle the callbacks then in some situations they don't get throttled 🤯