lua-api/process: Wait on child in gc
Pinnacle stores a receiver for the exit status of all spawned processes in case the client calls `wait`, which needs to provide the exit status. If we don't call `wait`, we essentially create a memory leak because the receiver will never get dropped. Probably not the *best* design but my brain doesn't want to come up with something better apparently