justfile: use luarocks path --full as fallback to export LUA_{,C}PATH
When requesting paths with --full, luarocks returns the local path,
followed the content of the package.*path lua variable.

This commit make it so we invoke luarocks --full to get the correct
paths if available. If luarocks version is too old, we fallback on the
environment to emulate the --full (except this will not retrieve lua
standard package.path when the variable is not in env.)