layout: Update window floating loc on resize
The window have two 'properties' when it come to their location. One is
the surface location (i.e. where the surface is mapped in global space),
the other is the floating location (used to restore the surface when
switching to a floating mode).

Only the surface location was updated on a resize previously.