grab/move: Update window floating_loc during fullscreen move
problem: The previous commit preserve the location if the window
intersect the output's non-exclusive zone. This is done to prevent
moving a window if it wasn't moved across outputs. However, since
floating window are attached to the output containing the biggest area,
their position could be preserved, so they could be re-attached to the
previous output.

solution: update the floating_loc when moving a fullscreen or maximized
window.