resize: allow resize to use all edges
Previous behavior was checking in which quadrant the pointer was, which
only allowed corner resizing.

The new behavior is to split the area into a 3x3 grid, and find in which
cell the pointer falls into.
The center cell cannot really use that behavior. Instead we use some
vector math to determine which the edge to use.