xdg_shell: Respect client output request
This commit implement a way for an xdg_shell client to pick which output
it should be fullscreened on.

The requested output is only considered if the WlOutput in the request
is owned by a mapped output (meaning the Output has a defined geometry).
Otherwise it's discarded and the output has associated tags. Otherwise
the request is discarded and the current window output (or focused
output in case of unmapped window) will be used.