fix(xdg_popup): Don't send additional configure on reposition
Smithay PopupSurface::sent_reposition already does the full configure
sequence (xdg_popup::reposition > xdg_popup::configure >
xdg_surface::configure). The additional PopupSurface::send_configure()
tried to send a reactive reposition.

This behavior is invalid, since it's not allowed to send a reactive
reposition unless the client requested it, but a reposition can be
requested on a non-reactive popup. On top of that, the client would
receive 2 configure (one in response to the reposition request, the
other being a fake 'reaction').

While the server will not send the second configure sequence if unless
the positioner is marked as reactive, this is spamming the log, which is
aggravated by smithay only checking the cached positioner, which only
gets populated once the popup is assigned a buffer.