hack: Use focused output scale as fallback for initial fractional scale
So firefox 146 released recently, bringing with it fractional scale support. In classic firefox fashion, there are a few, let's just say, *quirks* with the implementation, notably regarding popups. Our current fractional scale implementation, for the most part, sends the fractional scale for the first output. If this is different from the scale of the output the popup will actually show up on, the popup will flicker at the wrong scale for a frame. Also, the positioning becomes completely thrown off. Doesn't happen on any other application, just firefox. This is a small hack to get that to stop.

Also that firefox update causes popups to not show up if I turn my monitor off and on, that's fun. No idea what's causing that. I've reproduced it on jay and cosmic but not on niri